+++ src/SoftTissue.Application/Controllers/LinearViscoelasticityController.cs
@@ -2,16 +2,14 @@
 using Microsoft.AspNetCore.Mvc;
 using SoftTissue.Application.Extensions;
 using SoftTissue.Core.Operations.LinearViscoelasticity.CalculateStrainSensitivityAnalysis.MaxwellModel;
-using SoftTissue.Core.Operations.LinearViscoelasticity.CalculateStressSensitivityAnalysis.MaxwellModel;
 using SoftTissue.Core.Operations.ViscoelasticModel.CalculateResults.Linear.CalculateStrain.MaxwellModel;
 using SoftTissue.Core.Operations.ViscoelasticModel.CalculateResults.Linear.CalculateStress.MaxwellModel;
-using SoftTissue.DataContract.LinearViscoelasticity.CalculateStrain;
-using SoftTissue.DataContract.LinearViscoelasticity.CalculateStress;
-using SoftTissue.DataContract.LinearViscoelasticity.CalculateStressSensitivityAnalysis;
-using SoftTissue.DataContract.ViscoelasticModel.CalculateResults.Linear;
+using SoftTissue.Core.Operations.ViscoelasticModel.CalculateResults.Linear.Maxwell;
+using SoftTissue.DataContract.ViscoelasticModel.CalculateResults;
 using SoftTissue.DataContract.ViscoelasticModel.CalculateResults.Linear.CalculateStrain.Maxwell;
 using SoftTissue.DataContract.ViscoelasticModel.CalculateResults.Linear.CalculateStress.Maxwell;
 using SoftTissue.DataContract.ViscoelasticModel.CalculateResults.Linear.Maxwell;
+using SoftTissue.DataContract.ViscoelasticModel.CalculateResultsSensitivityAnalysis;
 using SoftTissue.DataContract.ViscoelasticModel.CalculateResultsSensitivityAnalysis.Linear;
 using System.Threading.Tasks;
 
@@ -24,7 +22,7 @@ namespace SoftTissue.Application.Controllers
     public class LinearViscoelasticityController : ControllerBase
     {
         /// <summary>
-        /// It is responsible to execute an analysis to calculate the stress considering Maxwell Model.
+        /// It is responsible to calculate the stress for Maxwell Model.
         /// </summary>
         /// <param name="calculateMaxwellModelStress"></param>
         /// <param name="request"></param>
@@ -37,40 +35,17 @@ namespace SoftTissue.Application.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
-        [HttpPost("calculate-stress/maxwell-model")]
-        public async Task<ActionResult<CalculateStrainResponse>> CalculateStress(
+        [HttpPost("maxwell-model/calculate-stress")]
+        public async Task<ActionResult<CalculateResultsResponse>> CalculateStress(
             [FromServices] ICalculateMaxwellModelStress calculateMaxwellModelStress,
-            [FromQuery] CalculateMaxwellModelResultsRequest request)
+            [FromQuery] CalculateMaxwellModelStressRequest request)
         {
-            CalculateMaxwellModelResultsResponse response = await calculateMaxwellModelStress.ProcessAsync(request).ConfigureAwait(false);
+            CalculateResultsResponse response = await calculateMaxwellModelStress.ProcessAsync(request).ConfigureAwait(false);
             return response.BuildHttpResponse();
         }
 
         /// <summary>
-        /// It is responsible to execute a sensitivity analysis while calculating the stress considering Maxwell Model.
-        /// </summary>
-        /// <param name="calculateMaxwellModelStressSensitivityAnalysis"></param>
-        /// <param name="request"></param>
-        /// <returns></returns>
-        /// <response code="201">Returns the newly created files.</response>
-        /// <response code="400">If some validation do not passed.</response>
-        /// <response code="500">If occurred some error in process.</response>
-        /// <response code="501">If some resource is not implemented.</response>
-        [ProducesResponseType(StatusCodes.Status201Created)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        [ProducesResponseType(StatusCodes.Status501NotImplemented)]
-        [HttpPost("calculate-stress/maxwell-model/sensitivity-analysis")]
-        public async Task<ActionResult<CalculateStrainResponse>> CalculateStressSensitivityAnalysis(
-            [FromServices] ICalculateMaxwellModelStressSensitivityAnalysis calculateMaxwellModelStressSensitivityAnalysis,
-            [FromQuery] CalculateStressSensitivityAnalysisRequest request)
-        {
-            CalculateMaxwellModelResultsResponse response = await calculateMaxwellModelStressSensitivityAnalysis.ProcessAsync(request).ConfigureAwait(false);
-            return response.BuildHttpResponse();
-        }
-
-        /// <summary>
-        /// It is responsible to execute an analysis to calculate the strain considering Maxwell Model.
+        /// It is responsible to calculate the strain for Maxwell Model.
         /// </summary>
         /// <param name="calculateMaxwellModelStrain"></param>
         /// <param name="request"></param>
@@ -83,19 +58,19 @@ namespace SoftTissue.Application.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
-        [HttpPost("calculate-strain/maxwell-model")]
-        public async Task<ActionResult<CalculateStrainResponse>> CalculateStrain(
+        [HttpPost("maxwell-model/calculate-strain")]
+        public async Task<ActionResult<CalculateResultsResponse>> CalculateStrain(
             [FromServices] ICalculateMaxwellModelStrain calculateMaxwellModelStrain,
             [FromQuery] CalculateMaxwellModelStrainRequest request)
         {
-            CalculateStrainResponse response = await calculateMaxwellModelStrain.ProcessAsync(request).ConfigureAwait(false);
+            CalculateResultsResponse response = await calculateMaxwellModelStrain.ProcessAsync(request).ConfigureAwait(false);
             return response.BuildHttpResponse();
         }
 
         /// <summary>
-        /// It is responsible to execute a sensitivity analysis while calculating the strain considering Maxwell Model.
+        /// It is responsible to calculate the results for Maxwell Model.
         /// </summary>
-        /// <param name="calculateMaxwellModelStrainSensitivityAnalysis"></param>
+        /// <param name="calculateMaxwellModelResults"></param>
         /// <param name="request"></param>
         /// <returns></returns>
         /// <response code="201">Returns the newly created files.</response>
@@ -106,12 +81,35 @@ namespace SoftTissue.Application.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
-        [HttpPost("calculate-strain/maxwell-model/sensitivity-analysis")]
-        public async Task<ActionResult<CalculateStrainResponse>> CalculateStrain(
-            [FromServices] ICalculateMaxwellModelStrainSensitivityAnalysis calculateMaxwellModelStrainSensitivityAnalysis,
+        [HttpPost("maxwell-model/calculate-results")]
+        public async Task<ActionResult<CalculateResultsResponse>> CalculateResults(
+            [FromServices] ICalculateMaxwellModelResults calculateMaxwellModelResults,
+            [FromQuery] CalculateMaxwellModelResultsRequest request)
+        {
+            CalculateResultsResponse response = await calculateMaxwellModelResults.ProcessAsync(request).ConfigureAwait(false);
+            return response.BuildHttpResponse();
+        }
+
+        /// <summary>
+        /// It is responsible to calculate the results for Maxwell Model and execute a sensitivity analysis.
+        /// </summary>
+        /// <param name="calculateMaxwellModelStressSensitivityAnalysis"></param>
+        /// <param name="request"></param>
+        /// <returns></returns>
+        /// <response code="201">Returns the newly created files.</response>
+        /// <response code="400">If some validation do not passed.</response>
+        /// <response code="500">If occurred some error in process.</response>
+        /// <response code="501">If some resource is not implemented.</response>
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status501NotImplemented)]
+        [HttpPost("maxwell-model/calculate-results/sensitivity-analysis")]
+        public async Task<ActionResult<CalculateResultsSensitivityAnalysisResponse>> CalculateStressSensitivityAnalysis(
+            [FromServices] ICalculateMaxwellModelResultsSensitivityAnalysis calculateMaxwellModelStressSensitivityAnalysis,
             [FromQuery] CalculateMaxwellModelResultsSensitivityAnalysisRequest request)
         {
-            CalculateStrainResponse response = await calculateMaxwellModelStrainSensitivityAnalysis.ProcessAsync(request).ConfigureAwait(false);
+            CalculateResultsSensitivityAnalysisResponse response = await calculateMaxwellModelStressSensitivityAnalysis.ProcessAsync(request).ConfigureAwait(false);
             return response.BuildHttpResponse();
         }
     }