+++ src/SoftTissue.Application/Controllers/QuasiLinearViscoelasticityController.cs
@@ -2,14 +2,15 @@
 using Microsoft.AspNetCore.Mvc;
 using SoftTissue.Application.Extensions;
 using SoftTissue.Core.Operations.QuasiLinearViscoelasticity.CalculateConvergenceTime;
-using SoftTissue.Core.Operations.QuasiLinearViscoelasticity.CalculateStress.FungModel.DisregardRampTime.ConsiderSimplifiedRelaxationFunction;
 using SoftTissue.Core.Operations.QuasiLinearViscoelasticity.GenerateDomain.FungModel;
 using SoftTissue.Core.Operations.ViscoelasticModel.CalculateResults.QuasiLinear.ConsiderRampTime.Fung;
 using SoftTissue.Core.Operations.ViscoelasticModel.CalculateResults.QuasiLinear.ConsiderRampTime.SimplifiedFung;
 using SoftTissue.Core.Operations.ViscoelasticModel.CalculateResults.QuasiLinear.DisregardRampTime.Fung;
+using SoftTissue.Core.Operations.ViscoelasticModel.CalculateResults.QuasiLinear.DisregardRampTime.SimplifiedFung;
+using SoftTissue.DataContract.Models;
 using SoftTissue.DataContract.QuasiLinearViscoelasticity.CalculateConvergenceTime;
-using SoftTissue.DataContract.QuasiLinearViscoelasticity.CalculateStress.Fung.DisregardRampTime.ConsiderSimplifiedRelaxationFunction;
 using SoftTissue.DataContract.QuasiLinearViscoelasticity.GenerateDomain;
+using SoftTissue.DataContract.ViscoelasticModel.CalculateResults;
 using SoftTissue.DataContract.ViscoelasticModel.CalculateResults.QuasiLinear.ConsiderRampTime.Fung;
 using SoftTissue.DataContract.ViscoelasticModel.CalculateResults.QuasiLinear.ConsiderRampTime.SimplifiedFung;
 using SoftTissue.DataContract.ViscoelasticModel.CalculateResults.QuasiLinear.DisregardRampTime.Fung;
@@ -25,7 +26,7 @@ namespace SoftTissue.Application.Controllers
     public class QuasiLinearViscoelasticityController : ControllerBase
     {
         /// <summary>
-        /// It is responsible to execute an analysis to calculate the stress applied to Fung Model considering ramp time and the Reduced Relaxation Function.
+        /// It is responsible to calculate the results for Fung Model considering ramp time.
         /// </summary>
         /// <param name="operation"></param>
         /// <param name="request"></param>
@@ -38,17 +39,17 @@ namespace SoftTissue.Application.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
-        [HttpPost("calculate-stress/fung-model/consider-ramp-time")]
-        public async Task<ActionResult<CalculateFungModelResultsConsiderRampTimeResponse>> CalculateFungModelStressConsiderRampTime(
+        [HttpPost("fung-model/consider-ramp-time/calculate-results")]
+        public async Task<ActionResult<CalculateResultsResponse>> CalculateFungModelResultsConsiderRampTime(
             [FromServices] ICalculateFungModelResultsConsiderRampTime operation,
             [FromBody] CalculateFungModelResultsConsiderRampTimeRequest request)
         {
-            CalculateFungModelResultsConsiderRampTimeResponse response = await operation.ProcessAsync(request).ConfigureAwait(false);
+            CalculateResultsResponse response = await operation.ProcessAsync(request).ConfigureAwait(false);
             return response.BuildHttpResponse();
         }
 
         /// <summary>
-        /// It is responsible to execute an analysis to calculate the stress applied to Fung Model disregarding ramp time and considering the Reduced Relaxation Function.
+        /// It is responsible to calculate the results for Fung Model disregarding ramp time.
         /// </summary>
         /// <param name="operation"></param>
         /// <param name="request"></param>
@@ -61,17 +62,17 @@ namespace SoftTissue.Application.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
-        [HttpPost("calculate-stress/fung-model/disregard-ramp-time")]
-        public async Task<ActionResult<CalculateFungModelResultsDisregardRampTimeResponse>> CalculateFungModelStressDisregardRampTime(
-            [FromServices] ICalculateFungModelStressDisregardRampTime operation,
+        [HttpPost("fung-model/disregard-ramp-time/calculate-results")]
+        public async Task<ActionResult<CalculateResultsResponse>> CalculateFungModelResultsDisregardRampTime(
+            [FromServices] ICalculateFungModelResultsDisregardRampTime operation,
             [FromBody] CalculateFungModelResultsDisregardRampTimeRequest request)
         {
-            CalculateFungModelResultsDisregardRampTimeResponse response = await operation.ProcessAsync(request).ConfigureAwait(false);
+            CalculateResultsResponse response = await operation.ProcessAsync(request).ConfigureAwait(false);
             return response.BuildHttpResponse();
         }
 
         /// <summary>
-        /// It is responsible to execute an analysis to calculate the stress applied to Fung Model disregarding ramp time and considering the Simplified Reduced Relaxation Function.
+        /// It is responsible to calculate the results for Simplified Fung Model considering ramp time.
         /// </summary>
         /// <param name="operation"></param>
         /// <param name="request"></param>
@@ -84,17 +85,17 @@ namespace SoftTissue.Application.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
-        [HttpPost("calculate-stress/simplified-fung-model/consider-ramp-time")]
-        public async Task<ActionResult<CalculateSimplifiedFungModelResultsConsiderRampTimeResponse>> CalculateSimplifiedFungModelStressConsiderRampTime(
+        [HttpPost("simplified-fung-model/consider-ramp-time/calculate-results")]
+        public async Task<ActionResult<CalculateResultsResponse>> CalculateSimplifiedFungModelResultsConsiderRampTime(
             [FromServices] ICalculateSimplifiedFungModelResultsConsiderRampTime operation,
             [FromBody] CalculateSimplifiedFungModelResultsConsiderRampTimeRequest request)
         {
-            CalculateSimplifiedFungModelResultsConsiderRampTimeResponse response = await operation.ProcessAsync(request).ConfigureAwait(false);
+            CalculateResultsResponse response = await operation.ProcessAsync(request).ConfigureAwait(false);
             return response.BuildHttpResponse();
         }
 
         /// <summary>
-        /// It is responsible to execute an analysis to calculate the stress applied to Fung Model disregarding ramp time and considering the Reduced Relaxation Function.
+        /// It is responsible to execute an analysis to calculate the results applied to Fung Model disregarding ramp time and considering the Reduced Relaxation Function.
         /// </summary>
         /// <param name="operation"></param>
         /// <param name="request"></param>
@@ -107,17 +108,17 @@ namespace SoftTissue.Application.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
-        [HttpPost("calculate-stress/simplified-fung-model/disregard-ramp-time")]
-        public async Task<ActionResult<CalculateSimplifiedFungModelResultsDisregardRampTimeResponse>> CalculateSimplifiedFungModelStressDisregardRampTime(
-            [FromServices] ICalculateSimplifiedFungModelStressDisregardRampTime operation,
+        [HttpPost("calculate-results/simplified-fung-model/disregard-ramp-time")]
+        public async Task<ActionResult<CalculateResultsResponse>> CalculateSimplifiedFungModelResultsDisregardRampTime(
+            [FromServices] ICalculateSimplifiedFungModelResultsDisregardRampTime operation,
             [FromBody] CalculateSimplifiedFungModelResultsDisregardRampTimeRequest request)
         {
-            CalculateSimplifiedFungModelResultsDisregardRampTimeResponse response = await operation.ProcessAsync(request).ConfigureAwait(false);
+            CalculateResultsResponse response = await operation.ProcessAsync(request).ConfigureAwait(false);
             return response.BuildHttpResponse();
         }
 
         /// <summary>
-        /// It is responsible to execute an analysis to calculate the stress considering Fung Model.
+        /// It is responsible to execute an analysis to calculate the results considering Fung Model.
         /// </summary>
         /// <param name="operation"></param>
         /// <param name="request"></param>
@@ -130,60 +131,15 @@ namespace SoftTissue.Application.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         [ProducesResponseType(StatusCodes.Status501NotImplemented)]
-        [HttpPost("calculate-stress/simplifiedfung-model/disregard-ramp-time/experimental-model")]
-        public async Task<ActionResult<CalculateSimplifiedFungModelResultsDisregardRampTimeResponse>> CalculateSimplifiedFungModelStressDisregardRampTimeToExperimentalModel(
-            [FromServices] ICalculateSimplifiedFungModelStressDisregardRampTime operation,
-            [FromBody] CalculateSimplifiedFungModelStressDisregardRampTimeToExperimentalModelRequest request)
+        [HttpPost("calculate-results/simplifiedfung-model/disregard-ramp-time/experimental-model")]
+        public async Task<ActionResult<CalculateResultsResponse>> CalculateSimplifiedFungModelResultsDisregardRampTimeToExperimentalModel(
+            [FromServices] ICalculateSimplifiedFungModelResultsDisregardRampTime operation,
+            [FromQuery] double timeStep, double finalTime, double strain, ExperimentalModel experimentalModel)
         {
-            CalculateSimplifiedFungModelResultsDisregardRampTimeResponse response = await operation.ProcessAsync(request.CreateQuasiLinearRequest()).ConfigureAwait(false);
+            CalculateResultsResponse response = await operation.ProcessAsync(
+                CalculateSimplifiedFungModelResultsDisregardRampTimeRequest.Create(timeStep, finalTime, strain, experimentalModel)).ConfigureAwait(false);
             return response.BuildHttpResponse();
         }
-
-        /// <summary>
-        /// It is responsible to execute a sensitivity analysis while calculating the stress considering Fung Model.
-        /// </summary>
-        /// <param name="operation"></param>
-        /// <param name="request"></param>
-        /// <returns></returns>
-        /// <response code="201">Returns the newly created files.</response>
-        /// <response code="400">If some validation do not passed.</response>
-        /// <response code="500">If occurred some error in process.</response>
-        /// <response code="501">If some resource is not implemented.</response>
-        //[ProducesResponseType(StatusCodes.Status201Created)]
-        //[ProducesResponseType(StatusCodes.Status400BadRequest)]
-        //[ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        //[ProducesResponseType(StatusCodes.Status501NotImplemented)]
-        //[HttpPost("calculate-stress/fung-model/sensitivity-analysis")]
-        //public async Task<ActionResult<CalculateStressResponse>> CalculateStressSensivityAnalysis(
-        //    [FromServices] ICalculateFungModelStressSentivityAnalysis operation,
-        //    [FromBody] CalculateStressSensitivityAnalysisRequest request)
-        //{
-        //    CalculateStressResponse response = await calculateFungModelStressSentivityAnalysis.Process(request).ConfigureAwait(false);
-        //    return response.BuildHttpResponse();
-        //}
-
-        /// <summary>
-        /// It is responsible to execute a sensitivity analysis while calculating the stress considering Fung Model.
-        /// </summary>
-        /// <param name="operation"></param>
-        /// <param name="request"></param>
-        /// <returns></returns>
-        /// <response code="201">Returns the newly created files.</response>
-        /// <response code="400">If some validation do not passed.</response>
-        /// <response code="500">If occurred some error in process.</response>
-        /// <response code="501">If some resource is not implemented.</response>
-        //[ProducesResponseType(StatusCodes.Status201Created)]
-        //[ProducesResponseType(StatusCodes.Status400BadRequest)]
-        //[ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        //[ProducesResponseType(StatusCodes.Status501NotImplemented)]
-        //[HttpPost("calculate-stress/fung-model/sensitivity-analysis/explicit")]
-        //public async Task<ActionResult<CalculateStressResponse>> CalculateStressSensivityAnalysisExplicit(
-        //    [FromServices] ICalculateFungModelStressSentivityAnalysis operation,
-        //    [FromBody] CalculateStressSensitivityAnalysisExplicitRequest request)
-        //{
-        //    CalculateStressResponse response = await calculateFungModelStressSentivityAnalysis.Process(request.MapToSensitivityAnalysis()).ConfigureAwait(false);
-        //    return response.BuildHttpResponse();
-        //}
 
         /// <summary>
         /// It is responsible to generate the domain for Fung Model parameters.
@@ -210,7 +166,7 @@ namespace SoftTissue.Application.Controllers
         }
 
         /// <summary>
-        /// It is responsible to calculate the convergence time when the stress is constant.
+        /// It is responsible to calculate the convergence time when the results is constant.
         /// </summary>
         /// <param name="operation"></param>
         /// <param name="request"></param>