+++ src/SoftTissue.Core/Operations/LinearViscoelasticity/CalculateStressSensitivityAnalysis/MaxwellModel/ICalculateMaxwellModelStressSensitivityAnalysis.cs
@@ -1,9 +1,0 @@
-﻿using SoftTissue.Core.Models.Viscoelasticity.Linear.Maxwell;
-
-namespace SoftTissue.Core.Operations.LinearViscoelasticity.CalculateStressSensitivityAnalysis.MaxwellModel
-{
-    /// <summary>
-    /// It is responsible to do a semsitivity analysis while calculating the stress to Maxwell model.
-    /// </summary>
-    public interface ICalculateMaxwellModelStressSensitivityAnalysis : ICalculateLinearViscosityStressSensitivityAnalysis<MaxwellModelInput> { }
-}