+++ src/SoftTissue.Core/Operations/ViscoelasticModel/CalculateResultsSensitivityAnalysis/Linear/Maxwell/ICalculateMaxwellModelResultsSensitivityAnalysis.cs
@@ -1,0 +1,11 @@
+﻿using SoftTissue.Core.Models.Viscoelasticity.Linear.Maxwell;
+using SoftTissue.Core.Operations.ViscoelasticModel.CalculateResultsSensitivityAnalysis.Linear;
+using SoftTissue.DataContract.ViscoelasticModel.CalculateResultsSensitivityAnalysis.Linear;
+
+namespace SoftTissue.Core.Operations.LinearViscoelasticity.CalculateStrainSensitivityAnalysis.MaxwellModel
+{
+    /// <summary>
+    /// It is responsible to calculate the results with sensitivity analysis to a linear viscoelastic model.
+    /// </summary>
+    public interface ICalculateMaxwellModelResultsSensitivityAnalysis : ICalculateLinearModelResultsSensitivityAnalysis<CalculateMaxwellModelResultsSensitivityAnalysisRequest, MaxwellModelInput, MaxwellModelResult> { }
+}